Add files via upload
Ask the user to input two decimal numbers. Print these two decimal numbers on the next line separated by a space based on input order.

Print the message "The first one is greater than the second one" in the next line if the first decimal number inputted is greater than the second one.